v0.5.0 Redesigned sorting to be purely visual
Instead of actually reordering the underlying Columns vector of the Schedule, which was bound to cause issues, there's a separate vector of sorted indices which is used when actually displaying the Schedule.
There might still be some bugs in this, but it was necessary for undo / redo to work more stably.
Also removed the "sorted" bool from Column, since it's quite useless.