Initial autosave improvements (read desc.)
Write autosaves to separate "fileName_auto" files.
Copy autosave file to base file if the file is successfully closed (currently when opening or creating a new file. eventually also when closing the program).
Added additional check in IO_Handler::deleteSchedule to see if the file was successfully deleted by filesystem.