Make more filter editing events functional
Fixed ScheduleColumn::removeFilterGroup() trying to remove a Filter instead.
Event for adding Filters.
Added as listeners Schedule functions to the correct events.
Removed FilterEditorSubGui's "editing" bool, replaced openGroupCreate() with createGroupAndEdit(). When a FilterGroup is created, it will instantly be opened to edit, therefore there is no need for a different state.
Added remove button to FilterEditorSubGui that removes the FilterGroup being edited.