Create TimeWrapper, ClockTimeWrapper and redesign DateContainer
TimeWrapper is a much more usable wrapper for dates that can convert itself to the terrible tm struct. It has some useful functions that DateContainer has / had. Unlike the tm struct, it works with 1-based values and actual year numbers, but also has 0-based functionality, if needed.
DateContainer, instead of containing a tm struct, now contains a TimeWrapper.
ClockTimeWrapper is a smaller class that only works with hours and minutes. It can be used as its own separate type but it's also contained in TimeWrapper to allow for date and time to be stored in one type.
Changed most of the code that previously used the tm struct to now use TimeWrapper.
Added a constructor from ClockTimeWrapper to TimeContainer.