Implement select option renaming, add base for option color changing
Added a rename button to displayed select options in ElementEditorSubGui that turn the option name into a name input box when clicked.
Created a SelectOption struct that is used instead of simply the option name. It contains the option's name and color (not yet implemented).
Improved SelectOptionsModification to use std::optional for values other than the first index.
Added BLF_SelectOption and BLF_SelectOptions to read / write column select options. This allowed for the mutable bool to be moved from the BLF_Column :).
Renamed SelectOptionChange to SelectOptionUpdateInfo to reduce similarity with SelectOptionModification.
Fixed 3 cases in SelectContainer::update() not having breaks...

Also added SelectColor enum and colors to be used for select option colors later.