Hide column sort and remove buttons when they aren't needed
The sort button is only visible when the column header is hovered or it has a sort direction applied.
The remove button is only visible when the column header is hovered.
Moved the remove button to the right.
Turned the remove button into an image button the same size as the sort button.