Implement pretty good resizing of columns
Usually, columns are resized in a fixed area by changing the size of one and moving others (which may leave empty space). If the total size of the columns goes out of the view area, a horizontal scrollbar is added.
Needed to manually edit imgui_tables.cpp to add a COLUMN_MIN_WIDTH_OVERRIDE that is used for the minimum column width (instead of the default 1.0 which is way too small).

Fixed the main schedule window having horizontal scroll with shift + scroll. It does not anymore.