Add interface styles (currently Light and Dark, not fully finished)
Added functionality to create any amount of interface styles and switch between them in the new Preferences menu tab.
NOTE: There is a slight issue in that whatever colors a style overrides must also be overridden by all other styles or they will remain that of the old style. This is mostly remedied if every new style is based on the Light or Dark style.
NOTE: The style selection is not saved because there is no preferences I/O system yet.
Made a function to automatically calculate an okay-looking active table cell color based on the current style.
Made Event::addListener take a const function reference so that temporaries can be passed.
Made ImGui demo window always included but only shown in debug builds.
Slightly modified the program's icon.