Ok-ish font scale settings,
In the Preferences menu the user can select a font scale (Small - 0.75x, Normal - 1x or Large (1.25x).
Added the font scale to InterfaceStyleHandler. Font scale is not a part of GuiStyles but rather a separate part of the interface style.
Made various things scale better when the font size is changed.
Removed a bunch of unneeded constants from gui_sizes namespace.
Made an empty date the same size as a normal date in the DateEditor.
Moved the clear button next to the date in DateEditor.
Added a table argument to displayColumnContext so it doesn't have to get it in a weird way.