Made the program autosave when on OS shutdown
Made a signal handler / listener in Program. It's actually not really needed right now but might be in the future.
Made Window::terminate() invoke the windowCloseEvent so that an autosave can be made before the program closes.
Made input event listeners get passed as const ref to allow for temporaries to be passed.