Improved / fixed column context menu opening
It was impossible to reorder columns because clicking them would just open the context menu! Now, the behaviour is slightly better.
The context menu is opened only if the column header was not dragged at all.
When the context menu is open and the header is clicked again, it only closes the popup and doesn't reopen it right away. If it's closed with right click, it will reopen though. I think this is okay behavior since you can't drag or do much of anything else with right click.