Add GuiTextures::exists()
It's a neat function that returns true if a texture is found and false if it isn't, also sets the given GLuint to the texture's ID if it is found or the missing texture if it isn't.
So it makes sense to call getOrLoad() if a texture is absolutely required so it prints an error message and to call exists() to get optional textures.