Improve WiX packaging process
Instead of using a very complicated way to *remove* the 'schedules' directory from the WiX files, it now just is not added.
For archive packages like ZIP, 7z, etc the 'schedules' directory is just added in a much simpler way.
This means 'fix_wix.py' can be way simpler, which is good because i don't really trust it.