Implement Linux notification actions
Added "Mark as done" button to mark the previous item (based on start time) as finished.
Added very basic progress display as text in the form of "finished/total items".
Added "default" action handler (clicking on the notification itself) that just invokes the event to pass it on.
NOTE: The added notify loop might have or cause issues that haven't been found yet.