Organise subguis
Create and use an EventPipe class. An EventPipe is similar to an Event, but Events can be added to it that, when invoked, will be "passed through" the EventPipe. This is useful for creating chains of events that pass arguments to a final destination.
Organise the code of ScheduleGui and MainMenuBarGui along with their subguis into subdirectories.
Remove unnecessary child classes of TextInputModalSubGui where it could be used directly instead.
Add getListenerCount() function to Event.