Reset table settings when a new schedule is created
This is so that table settings from a previously opened file don't carry over.
If a file is open, and another file is opened, the settings should generally be replaced already.
But a new file starts with no table settings of its own.
Don't know if this even works properly - it's months old.